add mobile navbar
