+++ src/Components/VideoPlayer/VideoPlayer.scss
@@ -50,7 +50,7 @@ video {
   }
 }
 
-@media screen and (max-width: 479px) {
+@media screen and (max-width: 950px) {
   .video__container {
     margin-top: 2rem;
     padding: 0;
