+++ src/Pages/GradingFirst/VideoFirst/VideoFirst.jsx
@@ -7,7 +7,7 @@ import {
 import { useEffect } from "react";
 import "./VideoFirst.scss";
 import { useGetVideosQuery } from "../../../app/store/middleware/videosApi.js";
-import ThemeNavBar from "../../../Components/ThemeNavBar/ThemeNavBarResponsive.jsx";
+import ThemeNavBar from "../../../Components/ThemeNavBar/ThemeNavBar.jsx";
 
 export default function VideoFirst() {
   const { data, error, isLoading } = useGetVideosQuery();
